Feat: Creating the listing screen instrumented test.
