+++ listing/implementation/src/main/kotlin/com/will/listing/implementation/presentation/composable/component/SearchSection.kt
@@ -20,6 +20,7 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.drawBehind
 import androidx.compose.ui.platform.LocalConfiguration
+import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.tooling.preview.Preview
@@ -60,6 +61,7 @@ internal fun SearchSection(
                 drawRect(animatedColor)
             }
             .padding(top = animatedPadding)
+            .testTag("search")
     ) {
 
         var text by rememberSaveable { mutableStateOf("") }
@@ -67,7 +69,8 @@ internal fun SearchSection(
         OutlinedTextField(
             modifier = Modifier
                 .fillMaxWidth()
-                .padding(16.dp),
+                .padding(16.dp)
+                .testTag("search-input"),
             value = text,
             shape = RoundedCornerShape(40.dp),
             colors = TextFieldDefaults.colors(
