--- listing/implementation/src/test/snapshots/images/com.will.listing.implementation.presentation.composable.component_ListingPagingContentTest_validateComponent.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:3d59deab6cd8494e1a451bf90bf3bd3b26dd7498f73229eeb9f14123109b3196
-size 20226
