+++ listing/implementation/src/test/snapshots/images/com.will.listing.implementation.presentation.composable_ListingScreenSnapshotTest_validateComponent.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:6568c1f60dbb852ad624efc179121a07d1ec7e00262d5e47e43faa779c7d8f42
+size 17358
